// src/webparts/dashboardDemandeur/models/isp-list.model.ts
export interface ISPList {
  ID: string;

  Title: string;

  Comment: string;
  CtgVacation: string;

  StartDate: string;

  EndDate: string;

  NbrDays: number;

  RequestStatus: string;

  VacationType: string;

  ReplacedBy: string;

  CreatedBy: any;

  RemainingDays: any;

  Author: any;
  AttachmentFiles: any;
  IdERP: any;
  TrancheHoraire: string;

}